feat: Add login and registration links to navbar
The navbar in the App component has been updated to include login and registration links. When the user is not logged in, the navbar displays "Register" and "Login" links. When the user is logged in, the navbar displays a welcome message with the user's name. This enhancement improves the navigation and user experience of the application.